*fixed capacities calculation *capacity deviation by constants instead of by percentage *rearranged functions.py to utils folder with separate logic
