+++ constants.py
@@ -1,15 +1,12 @@
 # add_operator_capacity_constraint_not_weekend
-MIN_CAPACITY_NOT_WEEKEND = 0.5
-MAX_CAPACITY_NOT_WEEKEND = 1.3
+MIN_CAPACITY_NOT_WEEKEND = 2
+MAX_CAPACITY_NOT_WEEKEND = 1.5
 
 # add_weekly_capactiy_constraint
 MAX_WEEKLY_CAPACITY = 0.5
 
 # add_operator_capacity_constraint_nights
 MAX_NIGHT_CAPACITY = 1.2
-
-# add_operator_capacity_constraint_weekend
-INCREASE_MAX_SOFASHIM = 1.01
 
 # add_operator_capacity_constraint_weekend
 INCREASE_MAX_SOFASHIM = 1.01
